CODEBASE: Merge TypeAssertion files
