+++ src/utils/EnumHelper.ts
@@ -2,7 +2,7 @@ import type { Member } from "../types";
 import type { NetscriptContext } from "../Netscript/APIWrapper";
 
 import * as allEnums from "../Enums";
-import { assertString } from "../Netscript/TypeAssertion";
+import { assertString } from "../utils/TypeAssertion";
 import { errorMessage } from "../Netscript/ErrorMessages";
 import { getRandomIntInclusive } from "./helpers/getRandomIntInclusive";
 import { getRecordValues } from "../Types/Record";
