+++ src/utils/TypeAssertion.ts
@@ -1,4 +1,110 @@
 import type { Unknownify } from "../types";
+import type { NetscriptContext } from "../Netscript/APIWrapper";
+import { errorMessage } from "../Netscript/ErrorMessages";
+
 // This function is empty because Unknownify<T> is a typesafe assertion on any object with no runtime checks needed.
 // eslint-disable-next-line @typescript-eslint/no-empty-function
 export function assertLoadingType<T extends object>(val: object): asserts val is Unknownify<T> {}
+
+export class TypeAssertionError extends Error {
+  friendlyType: string;
+
+  constructor(message: string, friendlyType: string, options?: ErrorOptions) {
+    super(message, options);
+    this.name = this.constructor.name;
+    this.friendlyType = friendlyType;
+  }
+}
+
+/** Function for providing custom error message to throw for a type assertion.
+ * @param v: Value to assert type of
+ * @param assertFn: Typechecking function to use for asserting type of v.
+ * @param msgFn: Function to use to generate an error message if an error is produced. */
+export function assert<T>(
+  v: unknown,
+  assertFn: (v: unknown) => asserts v is T,
+  msgFn: (type: string) => string,
+): asserts v is T {
+  try {
+    assertFn(v);
+  } catch (e) {
+    if (e instanceof TypeAssertionError) {
+      throw msgFn(e.friendlyType);
+    }
+    const type = typeof e === "string" ? e : "unknown";
+    throw msgFn(type);
+  }
+}
+
+/** Returns the friendlyType of v. arrays are "array" and null is "null". */
+function getFriendlyType(v: unknown): string {
+  return v === null ? "null" : Array.isArray(v) ? "array" : typeof v;
+}
+
+export function isObject(v: unknown): v is Record<string, unknown> {
+  return getFriendlyType(v) === "object";
+}
+
+/** For non-objects, and for array/null, throws an error with the friendlyType of v. */
+export function objectAssert(v: unknown): asserts v is Record<string, unknown> {
+  const type = getFriendlyType(v);
+  if (type !== "object") {
+    console.error("The value is not an object. Value:", v);
+    throw new TypeAssertionError(
+      `The value is not an object. Its type is ${type}. Its string value is ${String(v)}.`,
+      type,
+    );
+  }
+}
+
+/** For non-string, throws an error with the friendlyType of v. */
+export function stringAssert(v: unknown): asserts v is string {
+  const type = getFriendlyType(v);
+  if (type !== "string") {
+    console.error("The value is not a string. Value:", v);
+    throw new TypeAssertionError(`The value is not an string. Its type is ${type}.`, type);
+  }
+}
+
+/** For non-array, throws an error with the friendlyType of v. */
+export function arrayAssert(v: unknown): asserts v is unknown[] {
+  if (!Array.isArray(v)) {
+    console.error("The value is not an array. Value:", v);
+    const type = getFriendlyType(v);
+    throw new TypeAssertionError(`The value is not an array. Its type is ${type}.`, type);
+  }
+}
+
+const userFriendlyString = (v: unknown): string => {
+  const clip = (s: string): string => {
+    if (s.length > 15) return s.slice(0, 12) + "...";
+    return s;
+  };
+  if (typeof v === "number") return String(v);
+  if (typeof v === "string") {
+    if (v === "") return "empty string";
+    return `'${clip(v)}'`;
+  }
+  const json = JSON.stringify(v);
+  if (!json) return "???";
+  return `'${clip(json)}'`;
+};
+
+export const debugType = (v: unknown): string => {
+  if (v === null) return `Is null.`;
+  if (v === undefined) return "Is undefined.";
+  if (typeof v === "function") return "Is a function.";
+  return `Is of type '${typeof v}', value: ${userFriendlyString(v)}`;
+};
+
+/**
+ * This function should be used to assert strings provided by the player. It uses a specialized utility function that
+ * provides a stack trace pointing to the player's invalid caller.
+ */
+export function assertString(ctx: NetscriptContext, argName: string, v: unknown): asserts v is string {
+  if (typeof v !== "string") throw errorMessage(ctx, `${argName} expected to be a string. ${debugType(v)}`, "TYPE");
+}
+
+export function assertFunction(ctx: NetscriptContext, argName: string, v: unknown): asserts v is () => void {
+  if (typeof v !== "function") throw errorMessage(ctx, `${argName} expected to be a function ${debugType(v)}`, "TYPE");
+}
